+++ Source/ACE/Entity/Player.cs
@@ -2746,17 +2746,5 @@ namespace ACE.Entity
 
             moveToObjectChain.EnqueueChain();
         }
-
-        public void HandleActionWorldBroadcast(string message, ChatMessageType messageType)
-        {
-            ActionChain chain = new ActionChain();
-            chain.AddAction(this, () => DoWorldBroadcast(message, messageType));
-            chain.EnqueueChain();
-        }
-
-        public void DoWorldBroadcast(string message, ChatMessageType messageType)
-        {
-            CurrentLandblock.EnqueueBroadcastGlobalChat(message, messageType);
-        }
     }
 }
